requirements.txt
pillow~=7.2.0
numpy~=1.20.1
Pyfhel~=2.3.1
dataclasses~=0.6